requirements.txt
streamlit>=1.31.0
openai>=1.12.0
google-generativeai>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.6
ollama>=0.1.6
python-dotenv>=1.0.0
pydantic>=2.5.0
plotly>=5.18.0
pytest>=8.0.0
